Enable apple-clang (macOS) compilation (#3)
* Update `effective_error_probability/CMakeLists.txt`

This update fixes compilation on Apple machines and enables successful compilation using Apple's Clang.

* Update `secret_key_rate/CMakeLists.txt`

This fixes compilation on Apple machines. In particular, this enables successful compilation with Apple's Clang. Note the the use of `fmt` library due to the lack of `std::format` support in Apple's Clang.

* Use `fmt` library instead of `std::format`

This introduces the use of `fmt` library due preprocessor directive. Note that `std::format` is not supported on Apple's Clang.

* Replace MSVC-specific literals

The literals, namely `Ui16`, has been replaced with `static_cast<uint16_t>(...)` to be cross-platform compatible.

* Update `README.md` with Apple-specific info